🚀 Fixed deployment build error
